// resources/js/pages/Reports/Pendapatan/models.ts
export interface pendapatanSum {
    ket: string;
    jumlah: number;
    ppn: number;
    materai: number;
    total: number;
    pot_pph: number;
    bayar: number;
}

export interface pendapatanDetail {
    tenant_book_id: number;
    tgl: string;
    jenis: string;
    toko: string;
    ket: string;
    no: string;
    jumlah: number;
    ppn: number;
    materai: number;
    total: number;
    pot_pph: number;
    bayar: number;
}